Moved everything out of JsgmeFileInstaller
